Unify app and fetch tradable tokens
